pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-music-playlist-generator"
version = "0.1.0"
description = "Slack-first automation backend for AI music playlist generation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115,<1",
  "google-api-python-client>=2.154,<3",
  "google-auth-httplib2>=0.2,<1",
  "google-auth-oauthlib>=1.2,<2",
  "httpx>=0.28,<1",
  "Pillow>=10.4,<12",
  "pydantic>=2.9,<3",
  "pydantic-settings>=2.5,<3",
  "python-multipart>=0.0.9,<1",
  "sqlalchemy>=2.0,<3",
  "uvicorn>=0.32,<1",
]

[project.optional-dependencies]
browser = [
  "playwright>=1.52,<2",
]
dev = [
  "pytest>=8,<9",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.setuptools.packages.find]
include = ["app*"]
